undo e2e validation changes, move validation test to examples and wire into build
