+++ test/e2e/expectedfailures/empty-parameter-steps.yaml
@@ -11,7 +11,7 @@ spec:
         - - name: a
             template: whalesay
             arguments:
-              parameters: []
+              parameters:
         - - name: b
             template: whalesay
             arguments:
