+++ test/e2e/expectedfailures/failed-step-event.yaml
@@ -6,11 +6,11 @@ kind: Workflow
 metadata:
   generateName: failed-step-event-
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: exit
   templates:
     - name: exit
       container:
         image: argoproj/argosay:v2
-        args: [exit, '1']
+        args: [exit, 1]
