+++ test/e2e/expectedfailures/failed-step.yaml
@@ -6,7 +6,7 @@ kind: Workflow
 metadata:
   generateName: failed-step-
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: failed-step
   templates:
