+++ test/e2e/expectedfailures/oom-kill.yaml
@@ -6,7 +6,7 @@ spec:
   entrypoint: fork-bomb
   templates:
   - name: fork-bomb
-    container:
+    script:
       image: debian:9.5-slim
       command: [bash, -c]
       args: [":(){ : $@$@;};: :"]
