+++ test/e2e/expectedfailures/volumes-pvc-fail-event.yaml
@@ -5,7 +5,7 @@ kind: Workflow
 metadata:
   generateName: volumes-pvc-fail-event-
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: volumes-pvc-example
   volumeClaimTemplates:
