+++ test/e2e/functional/continue-on-failed.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: continue-on-fail
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: workflow-ignore
   parallelism: 2
