+++ test/e2e/functional/dag-depends.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   generateName: dag-diamond-
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: diamond
   templates:
