+++ test/e2e/functional/param-aggregation.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   generateName: parameter-aggregation-
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: parameter-aggregation
   templates:
