+++ test/e2e/functional/stop-terminate.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: stop-terminate
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: diamond
   onExit: exit
