+++ test/e2e/functional/success-event.yaml
@@ -6,7 +6,7 @@ kind: Workflow
 metadata:
   generateName: success-event-
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: exit
   templates:
