+++ test/e2e/functional/synchronization-mutex-tmpl-level.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: synchronization-tmpl-level-mutex
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: synchronization-tmpl-level-example-mutex
   templates:
