+++ test/e2e/functional/synchronization-mutex-wf-level-1.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: synchronization-wf-level-mutex-1
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: whalesay
   synchronization:
