+++ test/e2e/functional/synchronization-mutex-wf-level.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: synchronization-wf-level-mutex
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: whalesay
   synchronization:
