+++ test/e2e/smoke/artifact-passing.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   generateName: artifact-passing-
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: main
   templates:
