+++ test/e2e/smoke/basic-2.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: basic-2
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: run-workflow
   templates:
