+++ test/e2e/smoke/basic.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   generateName: basic-
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: run-workflow
   templates:
