+++ test/e2e/smoke/cluster-workflow-template-whalesay-template.yaml
@@ -3,7 +3,7 @@ kind: ClusterWorkflowTemplate
 metadata:
   name: cluster-workflow-template-whalesay-template
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: whalesay-template
   arguments:
