+++ test/e2e/smoke/hello-world-workflow-tmpl.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   generateName: workflow-template-hello-world-
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: whalesay
   templates:
