+++ test/e2e/smoke/runasnonroot-workflow.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   generateName: runasnonroot-
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: main
   securityContext:
