+++ test/e2e/smoke/workflow-template-whalesay-template.yaml
@@ -3,7 +3,7 @@ kind: WorkflowTemplate
 metadata:
   name: workflow-template-whalesay-template
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: whalesay-template
   onExit: on-exit
