+++ test/e2e/testdata/archive-strategies.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: archive-strategies
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: main
   templates:
