+++ test/e2e/testdata/basic-clusterworkflowtemplate.yaml
@@ -3,11 +3,11 @@ kind: ClusterWorkflowTemplate
 metadata:
   name: basic
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   workflowMetadata:
     labels:
-      argo-e2e: 'true'
+      argo-e2e: true
   entrypoint: main
   templates:
     - name: main
