+++ test/e2e/testdata/basic-cronworkflow.yaml
@@ -3,11 +3,11 @@ kind: CronWorkflow
 metadata:
   name: basic
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   workflowMetadata:
     labels:
-      argo-e2e: 'true'
+      argo-e2e: true
   schedule: "* * * * *"
   workflowSpec:
     entrypoint: main
