+++ test/e2e/testdata/basic-workflow.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: basic
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: main
   templates:
