+++ test/e2e/testdata/exit-1.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: exit-1
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: exit
   templates:
