+++ test/e2e/testdata/input-artifacts.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: input-artifacts
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: main
   templates:
