+++ test/e2e/testdata/node-suspend.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: node-suspend
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: node-suspend
   templates:
