+++ test/e2e/testdata/same-input-output-path-optional.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   generateName: optional-input-output-artifact-same-location-
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: plan
   templates:
