+++ test/e2e/testdata/semaphore-tmpl-level.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: semaphore-tmpl-level
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: semaphore-tmpl-level-example
   templates:
