+++ test/e2e/testdata/semaphore-wf-level.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: semaphore-wf-level
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: whalesay
   synchronization:
