+++ test/e2e/testdata/sleep-3s.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: sleep-3s
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: sleep-3s
   templates:
