+++ test/e2e/testdata/storage-limit.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: storage-quota-limit
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: wait
   volumeClaimTemplates:                 # define volume, same syntax as k8s Pod spec
@@ -19,5 +19,4 @@ spec:
   - name: wait
     script:
       image: argoproj/argosay:v2
-      source: ''
       args: [echo, ":) Hello Argo!"]
