+++ test/e2e/testdata/two-items.yaml
@@ -3,7 +3,7 @@ kind: Workflow
 metadata:
   name: two-items
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: main
   templates:
