+++ test/e2e/testdata/workflow-template-nested-template.yaml
@@ -3,7 +3,7 @@ kind: WorkflowTemplate
 metadata:
   name: workflow-template-nested-template
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   templates:
     - name: whalesay-inner-template
