+++ test/e2e/testdata/workflow-template-with-enum-values.yaml
@@ -3,7 +3,7 @@ kind: WorkflowTemplate
 metadata:
   name: workflow-template-with-enum-values
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   entrypoint: argosay
   arguments:
