+++ test/e2e/testdata/workflow-templates/invalid-workflowtemplate.yaml
@@ -3,11 +3,11 @@ kind: WorkflowTemplate
 metadata:
   name: basic
   labels:
-    argo-e2e: 'true'
+    argo-e2e: true
 spec:
   workflowMetadata:
     labels:
-      argo-e2e: 'true'
+      argo-e2e: true
   entrypoints: main
   templates:
     - name: main
